# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: NFC Pay Starter API
  version: "1.0.0"
servers:
  - url: https://{baseUrl}
    variables:
      baseUrl:
        default: example.netlify.app
paths:
...
  /api/sandbox/pay:
    post:
      summary: Sandbox - mark session as paid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId: { type: string }
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  status: { type: string }
        "302": { description: Redirect if form-submitted }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...